requirements.txt
langchain
PyPDF2
streamlit
openai
faiss-cpu
python-dotenv
tiktoken
nltk 
rouge-score
PyPDF2 
langchain-openai
